String Arithmetic/requirements.txt
absl-py==0.15.0
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
Brlapi==0.7.0
cachetools==5.2.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cupy-cuda111==11.3.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
easydict==1.9
entrypoints==0.3
executing==0.8.3
fasteners==0.14.1
fastrlock==0.8.1
flatbuffers==1.12
fonttools==4.30.0
future==0.18.2
gast==0.3.3
geographiclib==1.52
geopy==2.2.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h11==0.13.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
importlib-metadata==5.0.0
importlib-resources==5.4.0
install==1.3.5
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
netifaces==0.10.4
notebook==6.4.10
numpy==1.23.4
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.1.1
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
reportlab==3.5.34
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==4.1.5
Send2Trash==1.8.0
simplejson==3.16.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
ssh-import-id==5.10
stack-data==0.2.0
systemd-python==234
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.7
uvicorn==0.11.3
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==2.2.2
widgetsnbextension==3.6.0
wrapt==1.12.1
wsproto==1.1.0
xkit==0.0.0
zipp==3.7.0
